+++ src/ace_routine/CoroutineScheduler.h
@@ -29,6 +29,7 @@ SOFTWARE.
   #include <Arduino.h> // Serial, Print
 #endif
 #include "Coroutine.h"
+#include "CoroutineProfiler.h"
 
 class Print;
 
@@ -172,7 +173,13 @@ class CoroutineSchedulerTemplate {
           // its continuation context determines whether to call
           // Coroutine::isDelayExpired(), Coroutine::isDelayMicrosExpired(), or
           // Coroutine::isDelaySecondsExpired().
-          (*mCurrent)->runCoroutine();
+          //
+          // The `CoroutineScheduler` always enables the profiler by calling
+          // `Coroutine::runCoroutineWithProfiler()`. If memory usage is a
+          // problem, then consider calling the `Coroutine::runCoroutine()`
+          // directly in the global `loop()` function, instead of going through
+          // the `CoroutineScheduler`.
+          (*mCurrent)->runCoroutineWithProfiler();
           break;
 
         case T_COROUTINE::kStatusEnding:
@@ -188,7 +195,6 @@ class CoroutineSchedulerTemplate {
       // Go to the next coroutine
       mCurrent = (*mCurrent)->getNext();
     }
-
 
     /** List all the routines in the linked list to the printer. */
     void listCoroutines(Print& printer) {